post created

+++ database/migrations/2022_02_04_055302_adminstration_table.php
@@ -14,7 +14,7 @@ class AdminstrationTable extends Migration
     public function up()
     {
         Schema::create('administrations', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->bigInteger('id')->primary();
             $table->string('userName');
             $table->string('password');
             $table->string('rePassword');
